+++ private/buf/cmd/buf/buf.go
@@ -67,6 +67,9 @@ import (
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/plugin/pluginprune"
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/plugin/pluginpush"
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/plugin/pluginupdate"
+	"github.com/bufbuild/buf/private/buf/cmd/buf/command/policy/policyprune"
+	"github.com/bufbuild/buf/private/buf/cmd/buf/command/policy/policypush"
+	"github.com/bufbuild/buf/private/buf/cmd/buf/command/policy/policyupdate"
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/push"
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/registry/module/modulecommit/modulecommitaddlabel"
 	"github.com/bufbuild/buf/private/buf/cmd/buf/command/registry/module/modulecommit/modulecommitinfo"
@@ -207,6 +210,16 @@ func NewRootCommand(name string) *appcmd.Command {
 					pluginpush.NewCommand("push", builder),
 					pluginupdate.NewCommand("update", builder),
 					pluginprune.NewCommand("prune", builder),
+				},
+			},
+			{
+				Use:    "policy",
+				Short:  "Work with policies",
+				Hidden: true,
+				SubCommands: []*appcmd.Command{
+					policypush.NewCommand("push", builder),
+					policyupdate.NewCommand("update", builder),
+					policyprune.NewCommand("prune", builder),
 				},
 			},
 			{