+++ private/bufpkg/bufcheck/bufcheckserver/internal/buflintvalidate/buflintvalidate.go
@@ -35,6 +35,9 @@ func CheckMessage(
 	if err != nil {
 		return err
 	}
+	if messageRules == nil {
+		return nil
+	}
 	return checkCELForMessage(
 		addAnnotationFunc,
 		messageRules,