+++ private/pkg/normalpath/normalpath_unix.go
@@ -37,7 +37,8 @@ func NormalizeAndValidate(path string) (string, error) {
 		return "", NewError(path, errNotRelative)
 	}
 	// https://github.com/bufbuild/buf/issues/51
-	if strings.HasPrefix(normalizedPath, normalizedRelPathJumpContextPrefix) {
+	if strings.HasPrefix(normalizedPath, normalizedRelPathJumpContextPrefix) ||
+		normalizedPath == normalizedRelPathJumpContextPath {
 		return "", NewError(path, errOutsideContextDir)
 	}
 	return normalizedPath, nil