+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+- Update exclude types to remove unused options reducing the size of generated code.
+- Add `gitlab-code-quality` error format to print errors in the GitLab Code Quality format
+  for `buf lint` and `buf breaking`.
+
+## [v1.56.0] - 2025-07-31
+
 - Add `buf export --all` flag to include non-proto source files.
 - Add s390x binaries for Linux to releases.
 - Fix ppc64le binaries for Linux released as x86_64 binaries.
@@ -1395,7 +1401,8 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.55.1...HEAD
+[Unreleased]: https://github.com/bufbuild/buf/compare/v1.56.0...HEAD
+[v1.56.0]: https://github.com/bufbuild/buf/compare/v1.55.1...v1.56.0
 [v1.55.1]: https://github.com/bufbuild/buf/compare/v1.55.0...v1.55.1
 [v1.55.0]: https://github.com/bufbuild/buf/compare/v1.54.0...v1.55.0
 [v1.54.0]: https://github.com/bufbuild/buf/compare/v1.53.0...v1.54.0